backend/requirements.txt
fastapi
uvicorn
gitpython
pytest
google-generativeai
pytest-cov
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
